+++ blocks/generate-sourcemaps.js
@@ -1,6 +1,0 @@
-/**
- * @module generate-sourcemaps
- */
-module.exports = type => ({
-  devtool: type
-})